Examine card number groupings as a means of preventing false positives.
Premise:
A Visa card that appears in 4 groups of 4 digits e.g. 4111 1111 1111 1111 was probably a credit card number.
A Visa card that appears as a single block of 16 digits, 4111111111111111, we can't tell and must treat it as a credit card number.
But if a Visa card appears in a strange grouping such as 41 11 11 11 11 11 11 11, it probably is NOT a credit card number and we should avoid redacting it.