Allow cards followed by expiration to be redacted
Since https://github.com/zendesk/credit_card_sanitizer/pull/43, strings
of the form '<card_number><whitespace><expiration_date>' where
<expiration_date> matches regex EXPIRATION_DATE result in the card
number not being redacted. This is due to `without_expiration` wrapping
<expiration_date> (including leading whitespace) in a random string,
which causes `valid_postfix?` to fail due to
`!ALPHANUMERIC.match(postfix[0])`. As the postfix is not valid,
redaction does not occur.

This change moves the leading whitespace in <expiration_date> outside of
the wrapping, so that `valid_postfix?` will succeed.

This also changes the bottom level `before` block in the tests to create
a sanitizer with `parse_flanking: true`. This is due to tests on lines
188-201 which (it appears) should pass with `parse_flanking: true`.